+++ crash/src/crash.gd
@@ -10,12 +10,29 @@ const quotes: Array[String] = [
 	'if (!working) crash();',
 	'Space Breaker 2%',
 	'Should have used a good engine',
-	'haha godot crash handler go brrr',
-	'top 10 null object reference',
+	'Haha godot crash handler go brrr',
+	'Top 10 null object reference',
 	'if (usingLE) memoryLeak();',
 	'https://github.com/what-is-a-git/FunkinGodot',
-	'forgor a ;',
-	'play phighting'
+	'https://github.com/riconuts/fnf-troll-engine',
+	'Forgor a ;',
+	'Play Phighting',
+	'Happy with your decisions that led you to this point?',
+	'It\'s not too late, port your mod to something good!',
+	'Hey,\natleast, it\'s not VSlice!',
+	'Hey!\nDid you know, there are more than 6000 species of frogs?',
+	'Maybe download more ram?',
+	'In the name of Hades, Olympus, I accept this crash!',
+	'You didn\'t try and chart 18k, did you?',
+	'Guys help im running out of messages to put here',
+	'If you are seeing this,\nDM @vortex5694 on Discord the word \'watermelon\'',
+	'Hey! Did you know:\nHaxe is an open source high-level\nstrictly-typed programming language\nwith a fast optimizing cross-compiler.',
+	'I wonder if it would have been funnier\nif this crash handler was written in pure C...',
+	'Ignore the text below this. I\'m glad it crashed.',
+	'So John Shedletsky here, if you guys dont know\nthis is the guy from Roblox Forsaken.',
+	'OH MY GOD\nIT\'S DOOM SUCKER (from Forsaken)',
+	'Leather Engine (Crashes Offline)',
+	'Wow\n!\nthis\nmessage\nhas\nso\nmany\nlines\nha\nhaaa\nha\nhhaaa\nhaa\nhga\n\n\n\n\n\n\n\n\n\n\nha'
 ]
 
 @onready var root: Control = %root